+++ docs/changelog.rst
@@ -8,6 +8,8 @@ v2.0.0 (2019-XX-XX)
 
 - Massive rewrite of Starbelly's I/O to use Trio instead of asyncio.
 - Upgrade web client to Dart 2 and Angular 5.
+- Added repository hygiene checks to detect stray artifact files and unfinished
+  work markers in Python source/test modules.
 
 
 v1.0.0 (2017-11-03)